refactor: submit article modal hook
