+++ packages/shared/src/components/sidebar/ContributeSection.tsx
@@ -1,8 +1,8 @@
 import dynamic from 'next/dynamic';
-import { useRouter } from 'next/router';
-import React, { ReactElement, useContext, useEffect, useState } from 'react';
+import React, { ReactElement, useContext, useState } from 'react';
 import AnalyticsContext from '../../contexts/AnalyticsContext';
 import { FeaturesData } from '../../contexts/FeaturesContext';
+import { useSubmitArticle } from '../../hooks/useSubmitArticle';
 import EmbedIcon from '../icons/Embed';
 import LinkIcon from '../icons/Link';
 import { ListIcon, SidebarMenuItem } from './common';
@@ -35,9 +35,9 @@ export function ContributeSection({
   submitArticleModalButton,
   ...props
 }: SectionCommonProps & SubmitFlags): ReactElement {
-  const router = useRouter();
   const { trackEvent } = useContext(AnalyticsContext);
-  const [showSubmitArticle, setShowSubmitArticle] = useState(false);
+  const { isOpen: showSubmitArticle, onIsOpen: setShowSubmitArticle } =
+    useSubmitArticle();
   const [showNewSourceModal, setShowNewSourceModal] = useState(false);
   const contributeMenuItems: SidebarMenuItem[] = [
     {
@@ -70,19 +70,6 @@ export function ContributeSection({
     };
     contributeMenuItems.unshift(submitArticleMenuItem);
   }
-
-  useEffect(() => {
-    const search = new URLSearchParams(window.location.search);
-    const query = Object.fromEntries(search);
-    if (!query?.scout) {
-      return;
-    }
-
-    const { origin, pathname } = window.location;
-    setShowSubmitArticle(true);
-    router.replace(origin + pathname);
-  }, []);
-
   return (
     <>
       <Section
